shoshchat/requirements.txt
Django>=5.0
psycopg2-binary>=2.9
django-tenants>=3.5
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
python-decouple>=3.8
requests>=2.31
PyPDF2>=3.0
python-docx>=1.1
beautifulsoup4>=4.12
dj-stripe>=2.8
celery>=5.3
redis>=5.0
gunicorn>=21.2
